api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
pydantic[email]==2.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
